Extract job information before starting internal handling of MCDA methods
