+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/prioritization/topsis/TopsisMethod.java
@@ -19,19 +19,31 @@
 
 package org.planqk.nisq.analyzer.core.prioritization.topsis;
 
+import java.util.Objects;
+
 import org.planqk.nisq.analyzer.core.model.McdaJob;
+import org.planqk.nisq.analyzer.core.prioritization.JobDataExtractor;
+import org.planqk.nisq.analyzer.core.prioritization.JobInformation;
 import org.planqk.nisq.analyzer.core.prioritization.McdaMethod;
+import org.planqk.nisq.analyzer.core.repository.McdaJobRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
 
 /**
  * Service implementing the TOPSIS method to prioritize analysis results of the NISQ Analyzer.
  */
 @Service
+@RequiredArgsConstructor
 public class TopsisMethod implements McdaMethod {
 
     private final static Logger LOG = LoggerFactory.getLogger(TopsisMethod.class);
+
+    private final JobDataExtractor jobDataExtractor;
+
+    private final McdaJobRepository mcdaJobRepository;
 
     @Override
     public String getName() {
@@ -46,7 +58,17 @@ public class TopsisMethod implements McdaMethod {
     @Override
     public void executeMcdaMethod(McdaJob mcdaJob) {
         LOG.debug("Starting TOPSIS MCDA method to prioritize job with ID: {}", mcdaJob.getJobId());
+        JobInformation jobInformation = jobDataExtractor.getJobInformationFromUuid(mcdaJob.getJobId());
 
-        // TODO
+        // abort if job can not be found and therefore no information available
+        if (Objects.isNull(jobInformation)) {
+            LOG.error("Unable to retrieve information about job with ID: {}", mcdaJob.getJobId());
+            mcdaJob.setState("failed");
+            mcdaJob.setReady(true);
+            mcdaJobRepository.save(mcdaJob);
+            return;
+        }
+
+        // TODO: perform Topsis
     }
 }
